requirements.txt
playwright>=1.35.0
pytest>=7.0
pytest-playwright>=0.3.0
pytest-xdist
python-dotenv
